+++ python/fala/__init__.py
@@ -30,4 +30,4 @@ __all__ = [
     "sdk",
     "__version__",
 ]
-__version__ = "0.7.4"
+__version__ = "0.7.5"